Refactoring.

+++ src/map_generator_system.cpp
@@ -47,7 +47,7 @@ static void load_enemy(unsigned int enemy_index, const rapidjson::Document& json
 
 	Animation& enemy_animation = registry.emplace<Animation>(entity);
 	enemy_animation.max_frames = 4;
-	enemy_animation.state = enemy_state_to_animation_state.at(static_cast<int>(enemy_component.state));
+	enemy_animation.state = enemy_state_to_animation_state.at(static_cast<size_t>(enemy_component.state));
 	registry.emplace<RenderRequest>(entity,
 									enemy_type_textures.at(static_cast<int>(enemy_component.type)),
 									EFFECT_ASSET_ID::ENEMY,
